+++ jiyoontrip/src/app/components/purchase-detail/prompt/2.ui.txt
@@ -67,9 +67,10 @@
 
 ==============================================
 
-핵심요구사항) 아래 노드아이디 참고해서 gap마다 선 넣어서 구현할 것.
+기존히스토리-참고만) 아래 노드아이디 참고해서 gap마다 선 넣어서 구현할 것.
             -노드ID: 285:32017
 
-핵심요구사항) noInquiry 영역 아래 노드ID 참고해서 구현할 것.
+기존히스토리-참고만) noInquiry 영역 아래 노드ID 참고해서 구현할 것.
             -노드ID: 285:32055  
-        
+==============================================
+핵심요구사항) delete.icon 좌측에 edit.icon 넣을 것.
