+++ jiyoontrip/src/app/purchase/[id]/edit/page.tsx
@@ -1,0 +1,6 @@
+import PurchaseWrite from "@/app/components/purchase-write";
+
+export default function PurchaseEditPage() {
+  return <PurchaseWrite />;
+}
+
